refactor(target_chains/solana): allow importing instruction and program args
