+++ target_chains/solana/programs/pyth-price-publisher/src/instruction.rs
@@ -1,26 +1,11 @@
-use {
-    crate::{accounts, ensure},
-    solana_program::{
-        account_info::AccountInfo, program_error::ProgramError, program_memory::sol_memcmp,
-        pubkey::Pubkey, system_program,
-    },
-};
-
-mod initialize;
-mod initialize_publisher;
-mod submit_prices;
-
-pub use {
-    initialize::initialize, initialize_publisher::initialize_publisher,
-    submit_prices::submit_prices,
-};
+use bytemuck::{Pod, Zeroable};
 
 /// Seed used to derive the config account.
-const CONFIG_SEED: &str = "CONFIG";
+pub const CONFIG_SEED: &str = "CONFIG";
 
 /// Seed used to derive the associated buffer account that publishers can
 /// write their updates into.
-const PUBLISHER_CONFIG_SEED: &str = "PUBLISHER_CONFIG";
+pub const PUBLISHER_CONFIG_SEED: &str = "PUBLISHER_CONFIG";
 
 #[repr(u8)]
 pub enum Instruction {
@@ -40,114 +25,35 @@ pub enum Instruction {
     InitializePublisher,
 }
 
+#[cfg(feature = "solana-program")]
 impl Instruction {
-    pub fn parse(input: &[u8]) -> Result<Instruction, ProgramError> {
+    pub fn parse(input: &[u8]) -> Result<Instruction, solana_program::program_error::ProgramError> {
         match input.first() {
             Some(0) => Ok(Instruction::Initialize),
             Some(1) => Ok(Instruction::SubmitPrices),
             Some(2) => Ok(Instruction::InitializePublisher),
-            _ => Err(ProgramError::InvalidInstructionData),
+            _ => Err(solana_program::program_error::ProgramError::InvalidInstructionData),
         }
     }
 }
 
-fn validate_publisher<'a>(
-    account: Option<&AccountInfo<'a>>,
-) -> Result<AccountInfo<'a>, ProgramError> {
-    let publisher = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    ensure!(ProgramError::MissingRequiredSignature, publisher.is_signer);
-    ensure!(ProgramError::InvalidArgument, publisher.is_writable);
-    Ok(publisher)
+#[derive(Debug, Clone, Copy, Zeroable, Pod)]
+#[repr(C, packed)]
+pub struct InitializeArgs {
+    pub config_bump: u8,
+    pub authority: [u8; 32],
 }
 
-fn validate_system<'a>(account: Option<&AccountInfo<'a>>) -> Result<AccountInfo<'a>, ProgramError> {
-    let system = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    ensure!(
-        ProgramError::InvalidArgument,
-        pubkey_eq(system.key, &system_program::id())
-    );
-    Ok(system)
+#[derive(Debug, Clone, Copy, Zeroable, Pod)]
+#[repr(C, packed)]
+pub struct InitializePublisherArgs {
+    pub config_bump: u8,
+    pub publisher_config_bump: u8,
+    pub publisher: [u8; 32],
 }
 
-fn validate_payer<'a>(account: Option<&AccountInfo<'a>>) -> Result<AccountInfo<'a>, ProgramError> {
-    let payer = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    ensure!(ProgramError::MissingRequiredSignature, payer.is_signer);
-    ensure!(ProgramError::InvalidArgument, payer.is_writable);
-    Ok(payer)
-}
-
-fn pubkey_eq(a: &Pubkey, b: &Pubkey) -> bool {
-    sol_memcmp(a.as_ref(), b.as_ref(), 32) == 0
-}
-
-fn validate_config<'a>(
-    account: Option<&AccountInfo<'a>>,
-    bump: u8,
-    program_id: &Pubkey,
-    require_writable: bool,
-) -> Result<AccountInfo<'a>, ProgramError> {
-    let config = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    let config_pda = Pubkey::create_program_address(&[CONFIG_SEED.as_bytes(), &[bump]], program_id)
-        .map_err(|_| ProgramError::InvalidInstructionData)?;
-    ensure!(
-        ProgramError::InvalidArgument,
-        pubkey_eq(config.key, &config_pda)
-    );
-    if require_writable {
-        ensure!(ProgramError::InvalidArgument, config.is_writable);
-    }
-    Ok(config)
-}
-
-fn validate_authority<'a>(
-    account: Option<&AccountInfo<'a>>,
-    config: &AccountInfo<'a>,
-) -> Result<AccountInfo<'a>, ProgramError> {
-    let authority = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    ensure!(ProgramError::MissingRequiredSignature, authority.is_signer);
-    ensure!(ProgramError::InvalidArgument, authority.is_writable);
-    let config_data = config.data.borrow();
-    let config = accounts::config::read(*config_data)?;
-    ensure!(
-        ProgramError::MissingRequiredSignature,
-        authority.key.to_bytes() == config.authority
-    );
-    Ok(authority)
-}
-
-fn validate_publisher_config<'a>(
-    account: Option<&AccountInfo<'a>>,
-    bump: u8,
-    publisher: &Pubkey,
-    program_id: &Pubkey,
-    require_writable: bool,
-) -> Result<AccountInfo<'a>, ProgramError> {
-    let publisher_config = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    let publisher_config_pda = Pubkey::create_program_address(
-        &[
-            PUBLISHER_CONFIG_SEED.as_bytes(),
-            &publisher.to_bytes(),
-            &[bump],
-        ],
-        program_id,
-    )
-    .map_err(|_| ProgramError::InvalidInstructionData)?;
-    if require_writable {
-        ensure!(ProgramError::InvalidArgument, publisher_config.is_writable);
-    }
-    ensure!(
-        ProgramError::MissingRequiredSignature,
-        pubkey_eq(publisher_config.key, &publisher_config_pda)
-    );
-    Ok(publisher_config)
-}
-
-fn validate_buffer<'a>(
-    account: Option<&AccountInfo<'a>>,
-    program_id: &Pubkey,
-) -> Result<AccountInfo<'a>, ProgramError> {
-    let buffer = account.cloned().ok_or(ProgramError::NotEnoughAccountKeys)?;
-    ensure!(ProgramError::InvalidArgument, buffer.is_writable);
-    ensure!(ProgramError::IllegalOwner, buffer.owner == program_id);
-    Ok(buffer)
+#[derive(Debug, Clone, Copy, Zeroable, Pod)]
+#[repr(C, packed)]
+pub struct SubmitPricesArgsHeader {
+    pub publisher_config_bump: u8,
 }
